Order functionality completed
Added:
-Unity's new Input System Package
-Tool for creating orders
-OrderController (component for player used to giving orders to units)

Modified:
-"GOAPAgent.cs" Add a method to return agent to idle state (required to interupt current plan)

Removed:
-"ActionNewAction.cs" (was accientally commited last week)

+++ Assets/Scripts/AI/Core/GOAP/GAOP_Components/GOAP_Action.cs
@@ -10,7 +10,7 @@ namespace GOAP
 
         public GameObject target;
         public float minExecDist;
-        public float cost = 1.0f;
+        public byte cost = 1;
         [HideInInspector]
         public bool inRange = false;
